backend/requirements.txt
# CampusSync Edge — Backend Dependencies
fastapi>=0.110
uvicorn>=0.28
pydantic>=2.6
python-multipart>=0.0.9
PyJWT>=2.8
supabase>=2.4
httpx>=0.27
slowapi>=0.1.9
# ML / Data
numpy>=1.26
scikit-learn>=1.4
onnxruntime>=1.17
# Document parsing
pypdfium2>=4.26
python-docx>=1.1
# Google Generative AI
google-generativeai>=0.5
# LangChain (text splitting)
langchain-text-splitters>=0.2
# Testing
pytest>=8.0
